repr: introduce abstract data type (ADT) module
Introduce a repr module, adt, to house our various abstract data types.
This keeps things a bit more organized. Also move the strconv module to
the root, to match the public organization of the repr crate.
