+++ src/repr/src/adt/regex.rs
@@ -7,23 +7,27 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-use serde::{Deserialize, Serialize};
+//! Regex utilities.
+
 use std::cmp::Ordering;
 use std::hash::{Hash, Hasher};
 
-/// A regular expression object suitable for use in `repr::Datum`. The
-/// object has perhaps surprising comparison semantics as a result: two regexes
-/// are considered equal iff their string representation is identical. The
-/// [`PartialOrd`], [`Ord`], and [`Hash`] implementations are similarly based
-/// upon the string representation. This is not the natural equivalence relation
-/// for regexes: for example, the regexes `aa*` and `a+` define the same
-/// language, but would not compare as equal with this implementation of
+use serde::{Deserialize, Serialize};
+
+/// A regular expression object suitable for use in `repr::Datum`.
+///
+/// The object has perhaps surprising comparison semantics as a result: two
+/// regexes are considered equal iff their string representation is identical.
+/// The [`PartialOrd`], [`Ord`], and [`Hash`] implementations are similarly
+/// based upon the string representation. This is not the natural equivalence
+/// relation for regexes: for example, the regexes `aa*` and `a+` define the
+/// same language, but would not compare as equal with this implementation of
 /// [`PartialEq`].
 ///
 /// TODO(benesch): watch for a more efficient binary serialization for
 /// [`regex::Regex`] (https://github.com/rust-lang/regex/issues/258). The
-/// `serde_regex` crate serializes to a string and is forced to recompile
-/// the regex during deserialization.
+/// `serde_regex` crate serializes to a string and is forced to recompile the
+/// regex during deserialization.
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct Regex(#[serde(with = "serde_regex")] pub regex::Regex);
