+++ src/repr/src/scalar.rs
@@ -11,21 +11,15 @@ use std::fmt::{self, Write};
 use std::hash::{Hash, Hasher};
 
 use chrono::{DateTime, FixedOffset, NaiveDate, NaiveDateTime, NaiveTime, TimeZone, Utc};
-use failure::{bail, format_err};
+use failure::format_err;
 use ordered_float::OrderedFloat;
 use serde::{Deserialize, Serialize};
 
-use self::datetime::DateTimeField;
-use self::decimal::Significand;
+use crate::adt::decimal::Significand;
+use crate::adt::interval::Interval;
 use crate::{ColumnType, DatumDict, DatumList};
 
-pub mod datetime;
-pub mod decimal;
-pub mod jsonb;
-pub mod regex;
-pub mod strconv;
-
-/// A literal value.
+/// A single value.
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Hash, Ord, PartialOrd)]
 pub enum Datum<'a> {
     /// An unknown value.
@@ -42,17 +36,15 @@ pub enum Datum<'a> {
     Float32(OrderedFloat<f32>),
     /// A 64-bit floating point number.
     Float64(OrderedFloat<f64>),
-    /// A Date
+    /// A date.
     Date(NaiveDate),
-    /// A Time without a Date
+    /// A time.
     Time(NaiveTime),
-    /// A DateTime
+    /// A date and time, without a timezone.
     Timestamp(NaiveDateTime),
-    /// A time zone aware DateTime
+    /// A date and time, with a timezone.
     TimestampTz(DateTime<Utc>),
-    /// A span of time
-    ///
-    /// Either a concrete number of seconds, or an abstract number of Months
+    /// A span of time.
     Interval(Interval),
     /// An exact decimal number, possibly with a fractional component, with up
     /// to 38 digits of precision.
@@ -61,15 +53,19 @@ pub enum Datum<'a> {
     Bytes(&'a [u8]),
     /// A sequence of Unicode codepoints encoded as UTF-8.
     String(&'a str),
-    /// A heterogenous sequence of Datums
+    /// A homogeneous sequence of `Datum`s.
     List(DatumList<'a>),
-    /// A mapping from string keys to Datums,
+    /// A mapping from string keys to `Datum`s.
     Dict(DatumDict<'a>),
-    /// Json null does not behave like SQL null :'(
+    /// An unknown value within a JSON-typed `Datum`.
+    ///
+    /// This variant is distinct from [`Datum::Null`] as a null datum is
+    /// distinct from a non-null datum that contains the JSON value `null`.
     JsonNull,
 }
 
 impl<'a> Datum<'a> {
+    /// Reports whether this datum is null (i.e., is [`Datum::Null`]).
     pub fn is_null(&self) -> bool {
         match self {
             Datum::Null => true,
@@ -164,6 +160,11 @@ impl<'a> Datum<'a> {
         )))
     }
 
+    /// Unwraps the boolean value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::False`] or [`Datum::True`].
     pub fn unwrap_bool(&self) -> bool {
         match self {
             Datum::False => false,
@@ -172,6 +173,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the 32-bit integer value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Int32`].
     pub fn unwrap_int32(&self) -> i32 {
         match self {
             Datum::Int32(i) => *i,
@@ -179,6 +185,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the 64-bit integer value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Int64`].
     pub fn unwrap_int64(&self) -> i64 {
         match self {
             Datum::Int64(i) => *i,
@@ -200,6 +211,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the 32-bit floating-point value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Float32`].
     pub fn unwrap_float32(&self) -> f32 {
         match self {
             Datum::Float32(f) => f.into_inner(),
@@ -207,6 +223,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the 64-bit floating-point value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Float64`].
     pub fn unwrap_float64(&self) -> f64 {
         match self {
             Datum::Float64(f) => f.into_inner(),
@@ -214,6 +235,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the date value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Date`].
     pub fn unwrap_date(&self) -> chrono::NaiveDate {
         match self {
             Datum::Date(d) => *d,
@@ -221,6 +247,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the time value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Time`].
     pub fn unwrap_time(&self) -> chrono::NaiveTime {
         match self {
             Datum::Time(t) => *t,
@@ -228,6 +259,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the timestamp value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Timestamp`].
     pub fn unwrap_timestamp(&self) -> chrono::NaiveDateTime {
         match self {
             Datum::Timestamp(ts) => *ts,
@@ -235,6 +271,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the timestamptz value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::TimestampTz`].
     pub fn unwrap_timestamptz(&self) -> chrono::DateTime<Utc> {
         match self {
             Datum::TimestampTz(ts) => *ts,
@@ -242,6 +283,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the interval value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Interval`].
     pub fn unwrap_interval(&self) -> Interval {
         match self {
             Datum::Interval(iv) => *iv,
@@ -249,6 +295,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the decimal value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Decimal`].
     pub fn unwrap_decimal(&self) -> Significand {
         match self {
             Datum::Decimal(d) => *d,
@@ -256,6 +307,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the string value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::String`].
     pub fn unwrap_str(&self) -> &'a str {
         match self {
             Datum::String(s) => s,
@@ -263,6 +319,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the bytes value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Bytes`].
     pub fn unwrap_bytes(&self) -> &'a [u8] {
         match self {
             Datum::Bytes(b) => b,
@@ -270,6 +331,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the list value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::List`].
     pub fn unwrap_list(&self) -> DatumList<'a> {
         match self {
             Datum::List(list) => *list,
@@ -277,6 +343,11 @@ impl<'a> Datum<'a> {
         }
     }
 
+    /// Unwraps the dict value within this datum.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the datum is not [`Datum::Dict`].
     pub fn unwrap_dict(&self) -> DatumDict<'a> {
         match self {
             Datum::Dict(dict) => *dict,
@@ -288,6 +359,7 @@ impl<'a> Datum<'a> {
         Datum::TimestampTz(DateTime::<Utc>::from_utc(ndt, Utc))
     }
 
+    /// Reports whether this datum is an instance of the specified column type.
     pub fn is_instance_of(self, column_type: &ColumnType) -> bool {
         fn is_instance_of_scalar(datum: Datum, scalar_type: &ScalarType) -> bool {
             if let ScalarType::Jsonb = scalar_type {
@@ -546,12 +618,10 @@ impl fmt::Display for Datum<'_> {
     }
 }
 
-/// The fundamental type of a [`Datum`].
+/// The type of a [`Datum`].
 ///
-/// A fundamental type is what is typically thought of as a type, like "Int32"
-/// or "String." The full [`ColumnType`] struct bundles additional information, like
-/// an optional default value and nullability, that must also be considered part
-/// of a datum's type.
+/// There is a direct correspondence between `Datum` variants and `ScalarType`
+/// variants.
 #[derive(Clone, Debug, Eq, Serialize, Deserialize, Ord, PartialOrd)]
 pub enum ScalarType {
     /// The type of an unknown datum. Whenever possible, this variant should be
@@ -561,37 +631,65 @@ pub enum ScalarType {
     /// `SELECT NULL`, where there is no additional hint about what type `NULL`
     /// should take on.
     Unknown,
+    /// The type of [`Datum::True`] and [`Datum::False`].
     Bool,
+    /// The type of [`Datum::Int32`].
     Int32,
+    /// The type of [`Datum::Int64`].
     Int64,
+    /// The type of [`Datum::Float32`].
     Float32,
+    /// The type of [`Datum::Float64`].
     Float64,
-    /// An exact decimal number with a specified precision and scale. The
-    /// precision constrains the total number of digits in the number, while the
-    /// scale specifies the number of digits after the decimal point. The
-    /// maximum precision is [`decimal::MAX_DECIMAL_PRECISION`]. The scale must
-    /// be less than or equal to the precision.
-    Decimal(u8, u8),
-    Date,
-    Time,
-    Timestamp,
-    TimestampTz,
-    /// A possibly-negative time span
+    /// The type of [`Datum::Decimal`].
     ///
-    /// Represented by the [`Interval`] enum
+    /// This type additionally specifies the precision and scale of the decimal
+    /// . The precision constrains the total number of digits in the number,
+    /// while the scale specifies the number of digits after the decimal point.
+    /// The maximum precision is [`MAX_DECIMAL_PRECISION`]. The scale
+    /// must be less than or equal to the precision.
+    ///
+    /// [`MAX_DECIMAL_PRECISION`]: crate::adt::decimal::MAX_DECIMAL_PRECISION
+    Decimal(u8, u8),
+    /// The type of [`Datum::Date`].
+    Date,
+    /// The type of [`Datum::Time`].
+    Time,
+    /// The type of [`Datum::Timestamp`].
+    Timestamp,
+    /// The type of [`Datum::TimestampTz`].
+    TimestampTz,
+    /// The type of [`Datum::Interval`].
     Interval,
+    /// The type of [`Datum::Bytes`].
     Bytes,
+    /// The type of [`Datum::String`].
     String,
-    /// Json behaves like postgres' jsonb type but is stored as Datum::JsonNull/True/False/String/Float64/List/Dict.
-    /// The sql type system is responsible for preventing these being used as normal sql datums without casting.
+    /// The type of a datum that may represent any valid JSON value.
+    ///
+    /// Valid datum variants for this type are:
+    ///
+    ///   * [`Datum::Null`]
+    ///   * [`Datum::False`]
+    ///   * [`Datum::True`]
+    ///   * [`Datum::String`]
+    ///   * [`Datum::Float64`]
+    ///   * [`Datum::List`]
+    ///   * [`Datum::Dict`]
     Jsonb,
-    /// A sensible array type (named List to leave the door open for supporting the postgres array type as Array)
-    /// Backed by a DatumList
-    /// Due to limits of sql type syntax, List elements are always allowed to be null
+    /// The type of [`Datum::List`].
+    ///
+    /// Elements within the list are of the specified type. List elements may
+    /// always be [`Datum::Null`].
     List(Box<ScalarType>),
 }
 
 impl<'a> ScalarType {
+    /// Returns the contained decimal precision and scale.
+    ///
+    /// # Panics
+    ///
+    /// Panics if the scalar type is not [`ScalarType::Decimal`].
     pub fn unwrap_decimal_parts(&self) -> (u8, u8) {
         match self {
             ScalarType::Decimal(p, s) => (*p, *s),
@@ -599,6 +697,7 @@ impl<'a> ScalarType {
         }
     }
 
+    /// Constructs a dummy datum for this type.
     pub fn dummy_datum(&self) -> Datum<'a> {
         match self {
             ScalarType::Unknown => Datum::Null,
@@ -726,579 +825,6 @@ impl fmt::Display for ScalarType {
             String => f.write_str("string"),
             Jsonb => f.write_str("jsonb"),
             List(t) => write!(f, "{}[]", t),
-        }
-    }
-}
-
-impl fmt::Display for ColumnType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "{}{}",
-            self.scalar_type,
-            if self.nullable { "?" } else { "" }
-        )
-    }
-}
-
-/// An interval of time meant to express SQL intervals.
-///
-/// Obtained by parsing an `INTERVAL '<value>' <unit> [TO <precision>]`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Serialize, Hash, Deserialize)]
-pub struct Interval {
-    /// A possibly negative number of months for field types like `YEAR`
-    pub months: i32,
-    /// A timespan represented in nanoseconds.
-    ///
-    /// Irrespective of values, `duration` will not be carried over into
-    /// `months`.
-    pub duration: i128,
-}
-
-impl Default for Interval {
-    fn default() -> Self {
-        Self {
-            months: 0,
-            duration: 0,
-        }
-    }
-}
-
-impl std::ops::Neg for Interval {
-    type Output = Self;
-    fn neg(self) -> Self {
-        Self {
-            months: -self.months,
-            duration: -self.duration,
-        }
-    }
-}
-
-impl Interval {
-    /// Constructs a new `Interval` with the specified units of time.
-    ///
-    /// `nanos` in excess of `999_999_999` are carried over into seconds.
-    pub fn new(months: i32, seconds: i64, nanos: i64) -> Result<Interval, failure::Error> {
-        let i = Interval {
-            months,
-            duration: i128::from(seconds) * 1_000_000_000 + i128::from(nanos),
-        };
-        // Don't let our duration exceed Postgres' min/max for those same fields,
-        // equivalent to:
-        // ```
-        // SELECT INTERVAL '2147483647 days 2147483647 hours 59 minutes 59.999999 seconds';
-        // SELECT INTERVAL '-2147483647 days -2147483647 hours -59 minutes -59.999999 seconds';
-        // ```
-        if i.duration > 193_273_528_233_599_999_999_000
-            || i.duration < -193_273_528_233_599_999_999_000
-        {
-            bail!(
-                "exceeds min/max interval duration +/-(2147483647 days 2147483647 hours \
-                59 minutes 59.999999 seconds)"
-            )
-        } else {
-            Ok(i)
-        }
-    }
-
-    pub fn checked_add(&self, other: &Self) -> Option<Self> {
-        let months = match self.months.checked_add(other.months) {
-            Some(m) => m,
-            None => return None,
-        };
-        let seconds = match self.dur_as_secs().checked_add(other.dur_as_secs()) {
-            Some(s) => s,
-            None => return None,
-        };
-
-        match Self::new(
-            months,
-            seconds,
-            i64::from(self.nanoseconds() + other.nanoseconds()),
-        ) {
-            Ok(i) => Some(i),
-            Err(_) => None,
-        }
-    }
-
-    /// Returns the total number of whole seconds in the `Interval`'s duration.
-    pub fn dur_as_secs(&self) -> i64 {
-        (self.duration / 1_000_000_000) as i64
-    }
-
-    /// Computes the year part of the interval.
-    ///
-    /// The year part is the number of whole years in the interval. For example,
-    /// this function returns `3.0` for the interval `3 years 4 months`.
-    pub fn years(&self) -> f64 {
-        (self.months / 12) as f64
-    }
-
-    /// Computes the month part of the interval.
-    ///
-    /// The whole part is the number of whole months in the interval, modulo 12.
-    /// For example, this function returns `4.0` for the interval `3 years 4
-    /// months`.
-    pub fn months(&self) -> f64 {
-        (self.months % 12) as f64
-    }
-
-    /// Computes the day part of the interval.
-    ///
-    /// The day part is the number of whole days in the interval. For example,
-    /// this function returns `5.0` for the interval `5 days 4 hours 3 minutes
-    /// 2.1 seconds`.
-    pub fn days(&self) -> f64 {
-        (self.dur_as_secs() / (60 * 60 * 24)) as f64
-    }
-
-    /// Computes the hour part of the interval.
-    ///
-    /// The hour part is the number of whole hours in the interval, modulo 24.
-    /// For example, this function returns `4.0` for the interval `5 days 4
-    /// hours 3 minutes 2.1 seconds`.
-    pub fn hours(&self) -> f64 {
-        ((self.dur_as_secs() / (60 * 60)) % 24) as f64
-    }
-
-    /// Computes the minute part of the interval.
-    ///
-    /// The minute part is the number of whole minutes in the interval, modulo
-    /// 60. For example, this function returns `3.0` for the interval `5 days 4
-    /// hours 3 minutes 2.1 seconds`.
-    pub fn minutes(&self) -> f64 {
-        ((self.dur_as_secs() / 60) % 60) as f64
-    }
-
-    /// Computes the second part of the interval.
-    ///
-    /// The second part is the number of fractional seconds in the interval,
-    /// modulo 60.0.
-    pub fn seconds(&self) -> f64 {
-        let s = (self.dur_as_secs() % 60) as f64;
-        let ns = f64::from(self.nanoseconds()) / 1e9;
-        s + ns
-    }
-
-    /// Computes the nanosecond part of the interval.
-    pub fn nanoseconds(&self) -> i32 {
-        (self.duration % 1_000_000_000) as i32
-    }
-
-    /// Computes the total number of seconds in the interval.
-    pub fn as_seconds(&self) -> f64 {
-        (self.months as f64) * 60.0 * 60.0 * 24.0 * 30.0
-            + (self.dur_as_secs() as f64)
-            + f64::from(self.nanoseconds()) / 1e9
-    }
-
-    /// Truncate the "head" of the interval, removing all time units greater than `f`.
-    pub fn truncate_high_fields(&mut self, f: DateTimeField) {
-        match f {
-            DateTimeField::Year => {}
-            DateTimeField::Month => self.months %= 12,
-            DateTimeField::Day => self.months = 0,
-            hms => {
-                self.months = 0;
-                self.duration %= nanos_multiplier(hms.next_largest()) as i128
-            }
-        }
-    }
-
-    /// Converts this `Interval`'s duration into `chrono::Duration`.
-    pub fn duration_as_chrono(&self) -> chrono::Duration {
-        use chrono::Duration;
-        // This can be converted into a single call with
-        // https://github.com/chronotope/chrono/pull/426
-        Duration::seconds(self.dur_as_secs() as i64)
-            + Duration::nanoseconds(self.nanoseconds() as i64)
-    }
-
-    /// Truncate the "tail" of the interval, removing all time units less than `f`.
-    /// # Arguments
-    /// - `f`: Round the interval down to the specified time unit.
-    /// - `fsec_max_precision`: If `Some(x)`, keep only `x` places of nanosecond precision.
-    ///    Must be `(0,6)`.
-    ///
-    /// # Errors
-    /// - If `fsec_max_precision` is not None or within (0,6).
-    pub fn truncate_low_fields(
-        &mut self,
-        f: DateTimeField,
-        fsec_max_precision: Option<u64>,
-    ) -> Result<(), failure::Error> {
-        use DateTimeField::*;
-        match f {
-            Year => {
-                self.months -= self.months % 12;
-                self.duration = 0;
-            }
-            Month => {
-                self.duration = 0;
-            }
-            // Round nanoseconds.
-            Second => {
-                let default_precision = 6;
-                let precision = match fsec_max_precision {
-                    Some(p) => p,
-                    None => default_precision,
-                };
-
-                if precision > default_precision {
-                    bail!(
-                        "SECOND precision must be (0, 6), have SECOND({})",
-                        precision
-                    )
-                }
-
-                // Check if value should round up to nearest fractional place.
-                let remainder = self.duration % 10_i128.pow(9 - precision as u32);
-                if remainder / 10_i128.pow(8 - precision as u32) > 4 {
-                    self.duration += 10_i128.pow(9 - precision as u32);
-                }
-
-                self.duration -= remainder;
-            }
-            dhm => {
-                self.duration -= self.duration % nanos_multiplier(dhm) as i128;
-            }
-        }
-        Ok(())
-    }
-}
-
-/// Returns the number of nanoseconds in a single unit of `field`.
-pub fn nanos_multiplier(field: DateTimeField) -> i64 {
-    seconds_multiplier(field) * 1_000_000_000
-}
-
-/// Returns the number of seconds in a single unit of `field`.
-pub fn seconds_multiplier(field: DateTimeField) -> i64 {
-    use DateTimeField::*;
-    match field {
-        Day => 60 * 60 * 24,
-        Hour => 60 * 60,
-        Minute => 60,
-        Second => 1,
-        _other => unreachable!("Do not call with a non-duration field"),
-    }
-}
-
-/// Format an interval in a human form
-///
-/// Example outputs:
-///
-/// * 1 year 2 months 5 days 03:04:00
-/// * -1 year +5 days +18:59:29.3
-/// * 00:00:00
-impl fmt::Display for Interval {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let mut months = self.months;
-        let neg_mos = months < 0;
-        months = months.abs();
-        let years = months / 12;
-        months %= 12;
-        let mut secs = self.dur_as_secs().abs();
-        let mut nanos = self.nanoseconds().abs();
-        let days = secs / (24 * 60 * 60);
-        secs %= 24 * 60 * 60;
-        let hours = secs / (60 * 60);
-        secs %= 60 * 60;
-        let minutes = secs / 60;
-        secs %= 60;
-
-        if years > 0 {
-            if neg_mos {
-                f.write_char('-')?;
-            }
-            write!(f, "{} year", years)?;
-            if years > 1 {
-                f.write_char('s')?;
-            }
-        }
-
-        if months > 0 {
-            if years != 0 {
-                f.write_char(' ')?;
-            }
-            if neg_mos {
-                f.write_char('-')?;
-            }
-            write!(f, "{} month", months)?;
-            if months > 1 {
-                f.write_char('s')?;
-            }
-        }
-
-        if days > 0 {
-            if years > 0 || months > 0 {
-                f.write_char(' ')?;
-            }
-            if self.duration < 0 {
-                f.write_char('-')?;
-            } else if neg_mos {
-                f.write_char('+')?;
-            }
-            write!(f, "{} day", days)?;
-            if days != 1 {
-                f.write_char('s')?;
-            }
-        }
-
-        let non_zero_hmsn = hours > 0 || minutes > 0 || secs > 0 || nanos > 0;
-
-        if (years == 0 && months == 0 && days == 0) || non_zero_hmsn {
-            if years > 0 || months > 0 || days > 0 {
-                f.write_char(' ')?;
-            }
-            if self.duration < 0 && non_zero_hmsn {
-                f.write_char('-')?;
-            } else if neg_mos {
-                f.write_char('+')?;
-            }
-            write!(f, "{:02}:{:02}:{:02}", hours, minutes, secs)?;
-            if nanos > 0 {
-                let mut width = 9;
-                while nanos % 10 == 0 {
-                    width -= 1;
-                    nanos /= 10;
-                }
-                write!(f, ".{:0width$}", nanos, width = width)?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use super::*;
-
-    #[test]
-    fn interval_fmt() {
-        fn mon(mon: i32) -> String {
-            Interval {
-                months: mon,
-                ..Default::default()
-            }
-            .to_string()
-        }
-
-        assert_eq!(mon(1), "1 month");
-        assert_eq!(mon(12), "1 year");
-        assert_eq!(mon(13), "1 year 1 month");
-        assert_eq!(mon(24), "2 years");
-        assert_eq!(mon(25), "2 years 1 month");
-        assert_eq!(mon(26), "2 years 2 months");
-
-        fn dur(d: i64) -> String {
-            Interval::new(0, d, 0).unwrap().to_string()
-        }
-        assert_eq!(&dur(86_400 * 2), "2 days");
-        assert_eq!(&dur(86_400 * 2 + 3_600 * 3), "2 days 03:00:00");
-        assert_eq!(
-            &dur(86_400 * 2 + 3_600 * 3 + 60 * 45 + 6),
-            "2 days 03:45:06"
-        );
-        assert_eq!(&dur(86_400 * 2 + 3_600 * 3 + 60 * 45), "2 days 03:45:00");
-        assert_eq!(&dur(86_400 * 2 + 6), "2 days 00:00:06");
-        assert_eq!(&dur(86_400 * 2 + 60 * 45 + 6), "2 days 00:45:06");
-        assert_eq!(&dur(86_400 * 2 + 3_600 * 3 + 6), "2 days 03:00:06");
-        assert_eq!(&dur(3_600 * 3 + 60 * 45 + 6), "03:45:06");
-        assert_eq!(&dur(3_600 * 3 + 6), "03:00:06");
-        assert_eq!(&dur(3_600 * 3), "03:00:00");
-        assert_eq!(&dur(60 * 45 + 6), "00:45:06");
-        assert_eq!(&dur(60 * 45), "00:45:00");
-        assert_eq!(&dur(6), "00:00:06");
-
-        assert_eq!(&dur(-(86_400 * 2 + 6)), "-2 days -00:00:06");
-        assert_eq!(&dur(-(86_400 * 2 + 60 * 45 + 6)), "-2 days -00:45:06");
-        assert_eq!(&dur(-(86_400 * 2 + 3_600 * 3 + 6)), "-2 days -03:00:06");
-        assert_eq!(&dur(-(3_600 * 3 + 60 * 45 + 6)), "-03:45:06");
-        assert_eq!(&dur(-(3_600 * 3 + 6)), "-03:00:06");
-        assert_eq!(&dur(-(3_600 * 3)), "-03:00:00");
-        assert_eq!(&dur(-(60 * 45 + 6)), "-00:45:06");
-        assert_eq!(&dur(-(60 * 45)), "-00:45:00");
-        assert_eq!(&dur(-6), "-00:00:06");
-
-        fn mon_dur(mon: i32, d: i64) -> String {
-            Interval::new(mon, d, 0).unwrap().to_string()
-        }
-        assert_eq!(&mon_dur(1, 86_400 * 2 + 6), "1 month 2 days 00:00:06");
-        assert_eq!(
-            &mon_dur(1, 86_400 * 2 + 60 * 45 + 6),
-            "1 month 2 days 00:45:06"
-        );
-        assert_eq!(
-            &mon_dur(1, 86_400 * 2 + 3_600 * 3 + 6),
-            "1 month 2 days 03:00:06"
-        );
-        assert_eq!(
-            &mon_dur(26, 3_600 * 3 + 60 * 45 + 6),
-            "2 years 2 months 03:45:06"
-        );
-        assert_eq!(&mon_dur(26, 3_600 * 3 + 6), "2 years 2 months 03:00:06");
-        assert_eq!(&mon_dur(26, 3_600 * 3), "2 years 2 months 03:00:00");
-        assert_eq!(&mon_dur(26, 60 * 45 + 6), "2 years 2 months 00:45:06");
-        assert_eq!(&mon_dur(26, 60 * 45), "2 years 2 months 00:45:00");
-        assert_eq!(&mon_dur(26, 6), "2 years 2 months 00:00:06");
-
-        assert_eq!(
-            &mon_dur(26, -(86_400 * 2 + 6)),
-            "2 years 2 months -2 days -00:00:06"
-        );
-        assert_eq!(
-            &mon_dur(26, -(86_400 * 2 + 60 * 45 + 6)),
-            "2 years 2 months -2 days -00:45:06"
-        );
-        assert_eq!(
-            &mon_dur(26, -(86_400 * 2 + 3_600 * 3 + 6)),
-            "2 years 2 months -2 days -03:00:06"
-        );
-        assert_eq!(
-            &mon_dur(26, -(3_600 * 3 + 60 * 45 + 6)),
-            "2 years 2 months -03:45:06"
-        );
-        assert_eq!(&mon_dur(26, -(3_600 * 3 + 6)), "2 years 2 months -03:00:06");
-        assert_eq!(&mon_dur(26, -(3_600 * 3)), "2 years 2 months -03:00:00");
-        assert_eq!(&mon_dur(26, -(60 * 45 + 6)), "2 years 2 months -00:45:06");
-        assert_eq!(&mon_dur(26, -(60 * 45)), "2 years 2 months -00:45:00");
-        assert_eq!(&mon_dur(26, -6), "2 years 2 months -00:00:06");
-
-        assert_eq!(&mon_dur(-1, 86_400 * 2 + 6), "-1 month +2 days +00:00:06");
-        assert_eq!(
-            &mon_dur(-1, 86_400 * 2 + 60 * 45 + 6),
-            "-1 month +2 days +00:45:06"
-        );
-        assert_eq!(
-            &mon_dur(-1, 86_400 * 2 + 3_600 * 3 + 6),
-            "-1 month +2 days +03:00:06"
-        );
-        assert_eq!(
-            &mon_dur(-26, 3_600 * 3 + 60 * 45 + 6),
-            "-2 years -2 months +03:45:06"
-        );
-        assert_eq!(&mon_dur(-26, 3_600 * 3 + 6), "-2 years -2 months +03:00:06");
-        assert_eq!(&mon_dur(-26, 3_600 * 3), "-2 years -2 months +03:00:00");
-        assert_eq!(&mon_dur(-26, 60 * 45 + 6), "-2 years -2 months +00:45:06");
-        assert_eq!(&mon_dur(-26, 60 * 45), "-2 years -2 months +00:45:00");
-        assert_eq!(&mon_dur(-26, 6), "-2 years -2 months +00:00:06");
-
-        assert_eq!(
-            &mon_dur(-26, -(86_400 * 2 + 6)),
-            "-2 years -2 months -2 days -00:00:06"
-        );
-        assert_eq!(
-            &mon_dur(-26, -(86_400 * 2 + 60 * 45 + 6)),
-            "-2 years -2 months -2 days -00:45:06"
-        );
-        assert_eq!(
-            &mon_dur(-26, -(86_400 * 2 + 3_600 * 3 + 6)),
-            "-2 years -2 months -2 days -03:00:06"
-        );
-        assert_eq!(
-            &mon_dur(-26, -(3_600 * 3 + 60 * 45 + 6)),
-            "-2 years -2 months -03:45:06"
-        );
-        assert_eq!(
-            &mon_dur(-26, -(3_600 * 3 + 6)),
-            "-2 years -2 months -03:00:06"
-        );
-        assert_eq!(&mon_dur(-26, -(3_600 * 3)), "-2 years -2 months -03:00:00");
-        assert_eq!(
-            &mon_dur(-26, -(60 * 45 + 6)),
-            "-2 years -2 months -00:45:06"
-        );
-        assert_eq!(&mon_dur(-26, -(60 * 45)), "-2 years -2 months -00:45:00");
-        assert_eq!(&mon_dur(-26, -6), "-2 years -2 months -00:00:06");
-    }
-
-    #[test]
-    fn test_interval_value_truncate_low_fields() {
-        use DateTimeField::*;
-
-        let mut test_cases = [
-            (Year, None, (321, 654_321, 321_000_000), (26 * 12, 0, 0)),
-            (Month, None, (321, 654_321, 321_000_000), (321, 0, 0)),
-            (
-                Day,
-                None,
-                (321, 654_321, 321_000_000),
-                (321, 7 * 60 * 60 * 24, 0), // months: 321, duration: 604800s, is_positive_dur: true
-            ),
-            (
-                Hour,
-                None,
-                (321, 654_321, 321_000_000),
-                (321, 181 * 60 * 60, 0),
-            ),
-            (
-                Minute,
-                None,
-                (321, 654_321, 321_000_000),
-                (321, 10905 * 60, 0),
-            ),
-            (
-                Second,
-                None,
-                (321, 654_321, 321_000_000),
-                (321, 654_321, 321_000_000),
-            ),
-            (
-                Second,
-                Some(1),
-                (321, 654_321, 321_000_000),
-                (321, 654_321, 300_000_000),
-            ),
-            (
-                Second,
-                Some(0),
-                (321, 654_321, 321_000_000),
-                (321, 654_321, 0),
-            ),
-        ];
-
-        for test in test_cases.iter_mut() {
-            let mut i = Interval::new((test.2).0, (test.2).1, (test.2).2).unwrap();
-            let j = Interval::new((test.3).0, (test.3).1, (test.3).2).unwrap();
-
-            i.truncate_low_fields(test.0, test.1).unwrap();
-
-            if i != j {
-                panic!(
-                "test_interval_value_truncate_low_fields failed on {} \n actual: {:?} \n expected: {:?}",
-                test.0, i, j
-            );
-            }
-        }
-    }
-
-    #[test]
-    fn test_interval_value_truncate_high_fields() {
-        use DateTimeField::*;
-
-        let mut test_cases = [
-            (Year, (321, 654_321), (321, 654_321)),
-            (Month, (321, 654_321), (9, 654_321)),
-            (Day, (321, 654_321), (0, 654_321)),
-            (Hour, (321, 654_321), (0, 654_321 % (60 * 60 * 24))),
-            (Minute, (321, 654_321), (0, 654_321 % (60 * 60))),
-            (Second, (321, 654_321), (0, 654_321 % 60)),
-        ];
-
-        for test in test_cases.iter_mut() {
-            let mut i = Interval::new((test.1).0, (test.1).1, 123).unwrap();
-            let j = Interval::new((test.2).0, (test.2).1, 123).unwrap();
-
-            i.truncate_high_fields(test.0);
-
-            if i != j {
-                panic!(
-                    "test_interval_value_truncate_high_fields failed on {} \n actual: {:?} \n expected: {:?}",
-                    test.0, i, j
-                );
-            }
         }
     }
 }
